# docs/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "JSON API behind the Project Divert mobile app and admin tooling.\n\nThis document is generated from the application source by `scripts/generate_openapi.py`: paths and methods come from the Flask URL map, authentication and roles from the `jwt_required` decorator, request fields from the body and query reads in each view, and response codes from the status codes each view returns. `tests/api/test_openapi_contract.py` fails if it falls out of date.\n\nResponse property lists are best-effort: they are read from the serialiser each endpoint uses, so they name the keys you can expect rather than constraining their types.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/dodsy4/projectdivert/blob/main/LICENSE"
    },
    "title": "Project Divert API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/waste-requests/{request_id}/compliance/documents/{document_id}/verify": {
      "post": {
        "description": "Requires a bearer token with role: `admin`. Error messages returned by this endpoint: \"Compliance document not found\"; \"Failed to update compliance document\"; \"Waste request not found\"; \"expires_at must be later than issued_at\"; \"metadata must be an object\"; \"status must be verified, rejected, or expired\".",
        "operationId": "api_admin_verify_waste_request_compliance_document",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "expires_at": {},
                  "metadata": {},
                  "notes": {},
                  "status": {}
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "document": {},
                    "previous_status": {},
                    "request_id": {},
                    "summary": {},
                    "updated": {}
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Verify waste request compliance document (admin)",
        "tags": [
          "Compliance"
        ]
      }
    },
...
    "/api/v1/waste-requests/{request_id}/compliance": {
      "get": {
        "description": "Requires a bearer token with role: `admin`, `customer`, `driver`. Error messages returned by this endpoint: \"Forbidden\"; \"Waste request not found\".",
        "operationId": "api_get_waste_request_compliance",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "documents": {},
                    "request_id": {},
                    "request_status": {},
                    "summary": {}
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get waste request compliance",
        "tags": [
          "Compliance"
        ]
      }
    },
    "/api/v1/waste-requests/{request_id}/compliance/documents": {
      "post": {
        "description": "Requires a bearer token with role: `admin`, `driver`. Error messages returned by this endpoint: \"Drivers can only upload collection evidence documents\"; \"Failed to create compliance document\"; \"Forbidden\"; \"Invalid document_type\"; \"Invalid status\"; \"Only admins can set non-submitted status\"; \"Waste request not found\"; \"expires_at must be later than issued_at\"; \"file_url is required\"; \"file_url is too long\"; \"metadata must be an object\".",
        "operationId": "api_create_waste_request_compliance_document",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "document_reference": {},
                  "document_type": {},
                  "expires_at": {},
                  "file_url": {},
                  "issued_at": {},
                  "metadata": {},
                  "notes": {},
                  "status": {}
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "document": {},
                    "request_id": {},
                    "summary": {}
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create waste request compliance document",
        "tags": [
          "Compliance"
        ]
      }
    },
    "/api/v1/waste-requests/{request_id}/compliance/uploads": {
      "post": {
        "description": "Requires a bearer token with role: `admin`, `driver`. Error messages returned by this endpoint: \"Drivers can only upload collection evidence documents\"; \"Failed to save compliance upload\"; \"Forbidden\"; \"Invalid document_type\"; \"Waste request not found\"; \"file is required\".",
        "operationId": "api_upload_waste_request_compliance_file",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "document_type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "document_type": {}
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "document_type": {},
                    "request_id": {},
                    "upload": {}
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Upload waste request compliance file",
        "tags": [
          "Compliance"
        ]
      }
    },
    "/api/v1/waste-requests/{request_id}/compliance/uploads/sign": {
      "post": {
        "description": "Requires a bearer token with role: `admin`, `driver`. Error messages returned by this endpoint: \"Drivers can only upload collection evidence documents\"; \"Failed to create signed upload\"; \"Forbidden\"; \"Invalid document_type\"; \"Signed uploads require S3 storage backend\"; \"Waste request not found\"; \"file_name is required\"; \"mime_type is required\".",
        "operationId": "api_sign_waste_request_compliance_upload",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "document_type": {},
                  "file_name": {},
                  "mime_type": {}
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "document_type": {},
                    "request_id": {}
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Sign waste request compliance upload",
        "tags": [
          "Compliance"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This deployment",
      "url": "/"
    }
  ],
...
}